# Conflicts:
#	CHANGES
#	Cartfile.resolved
#	Carthage/Checkouts/MMTabBarView/MMTabBarView/MMTabBarView/MMTabBarView.h
#	Vienna.xcodeproj/project.pbxproj
#	Vienna/Sources/Alerts/NewSubscription.m
#	Vienna/Sources/Preferences window/SyncingPreferencesViewController.m
#	Vienna/Sources/Shared/SubscriptionModel.m
#	notes.html

+++ Vienna/Sources/Database/Import.m
@@ -62,7 +62,7 @@
         NSAlert *alert = [NSAlert new];
         alert.alertStyle = NSAlertStyleInformational;
         alert.messageText = NSLocalizedString(@"Import Completed", nil);
-        alert.informativeText = [NSString stringWithFormat:NSLocalizedString(@"%d subscriptions successfully imported", nil), countImported];
+        alert.informativeText = [NSString stringWithFormat:NSLocalizedString(@"%d subscriptions successfully imported", nil), (int)countImported];
         [alert runModal];
     }
 }
